# ByteString and AsciiString hashCode updates