--- a/microbench/src/main/java/io/netty/microbench/internal/PlatformDependentBenchmark.java
+++ b/microbench/src/main/java/io/netty/microbench/internal/PlatformDependentBenchmark.java
@@ -15,8 +15,11 @@
  */
 package io.netty.microbench.internal;
 
+import static io.netty.util.internal.StringUtil.asciiToLowerCase;
 import io.netty.microbench.util.AbstractMicrobenchmark;
+import io.netty.util.internal.HashCodeGenerator;
 import io.netty.util.internal.PlatformDependent;
+import io.netty.util.internal.StringUtil;
 
 import java.util.Arrays;
 
@@ -34,18 +37,25 @@ import org.openjdk.jmh.annotations.Threads;
 @State(Scope.Benchmark)
 public class PlatformDependentBenchmark extends AbstractMicrobenchmark {
 
-    @Param({ "10", "50", "100", "1000", "10000", "100000" })
+    @Param({ "10", "30", "50", "100", "1000", "10000", "100000" })
     private int size;
     private byte[] bytes1;
     private byte[] bytes2;
+    private char[] chars1;
+    private String string1;
+    private HashCodeGenerator hasher = PlatformDependent.hashCodeGenerator();
+    private HashCodeGenerator caseHasher = PlatformDependent.hashCodeGeneratorAsciiCaseInsensitive();
 
     @Setup(Level.Trial)
     public void setup() {
         bytes1 = new byte[size];
         bytes2 = new byte[size];
+        chars1 = new char[size];
         for (int i = 0; i < size; i++) {
             bytes1[i] = bytes2[i] = (byte) i;
+            chars1[i] = (char) i;
         }
+        string1 = new String(chars1);
     }
 
     @Benchmark
@@ -58,5 +68,75 @@ public class PlatformDependentBenchmark extends AbstractMicrobenchmark {
     @BenchmarkMode(Mode.Throughput)
     public boolean arraysBytesEqual() {
         return Arrays.equals(bytes1, bytes2);
+    }
+
+    @Benchmark
+    @BenchmarkMode(Mode.Throughput)
+    public int unsafeHashBytes() {
+        return hasher.hashCode(bytes1);
+    }
+
+    @Benchmark
+    @BenchmarkMode(Mode.Throughput)
+    public int oldHashBytes() {
+        int h = 0;
+        for (int i = 0; i < bytes1.length; ++i) {
+            h = h * 31 ^ bytes1[i] & 31;
+        }
+        return h;
+    }
+
+    @Benchmark
+    @BenchmarkMode(Mode.Throughput)
+    public int unsafeHashString() {
+        return hasher.hashCodeAsBytes(string1, 0, string1.length());
+    }
+
+    @Benchmark
+    @BenchmarkMode(Mode.Throughput)
+    public int oldHashString() {
+        int h = 0;
+        for (int i = 0; i < string1.length(); ++i) {
+            h = h * 31 ^ (byte) string1.charAt(i) & 31;
+        }
+        return h;
+    }
+
+    @Benchmark
+    @BenchmarkMode(Mode.Throughput)
+    public int unsafeHashCaseInsensitiveBytes() {
+        return caseHasher.hashCode(bytes1, 0, bytes1.length);
+    }
+
+    @Benchmark
+    @BenchmarkMode(Mode.Throughput)
+    public int oldHashCaseInsensitiveBytes() {
+        int h = 0;
+        for (int i = 0; i < bytes1.length; ++i) {
+            h = h * 31 ^ asciiToLowerCase(bytes1[i]) & 31;
+        }
+        return h;
+    }
+
+    @Benchmark
+    @BenchmarkMode(Mode.Throughput)
+    public int unsafeHashCaseInsensitiveString() {
+        return caseHasher.hashCodeAsBytes(string1, 0, string1.length());
+    }
+
+    @Benchmark
+    @BenchmarkMode(Mode.Throughput)
+    public int oldHashCaseInsensitiveString() {
+        int h = 0;
+        for (int i = 0; i < string1.length(); ++i) {
+            h = h * 31 ^ StringUtil.asciiToLowerCase((byte) string1.charAt(i)) & 31;
+        }
+        return h;
+    }
+
+    @Benchmark
+    @BenchmarkMode(Mode.Throughput)
+    public int arraysBytesHashCode() {
+        return Arrays.hashCode(bytes1);
     }
 }