… common/src/main/java/io/netty/util/internal/PlatformDependent0.java
… common/src/test/java/io/netty/util/AsciiStringTest.java
… common/src/test/java/io/netty/util/internal/PlatformDependentTest.java
… license/LICENSE.fnv1a.txt
… microbench/src/main/java/io/netty/microbenchmark/common/AsciiStringBenchmark.java