Motivation:
The ByteString class is often used in equality and hashCode operations. ByteString caches the hashCode after computed but case insensitive AsciiString hashCode operations are not cached.  The hash code algorithm we currently have is home grown and applies a mask of 31 onto every byte or character while generating the hash code which reduces the amount of bits each character contributes to the hash code from 8 down to 5. The JDK Arrays.hashCode(..) method does not support generating a hash code over a sub-section of an array which is required by ByteString and AsciiString.

Modifications:
- Use a hash code algorithm which takes all the bits of each character into account.
- Provide an optimized UNSAFE version of this algorithm which can work for case sensitive and insensitive.
- Update ByteString and AsciiString to take advantage of the new hashCode implementations.

Result:
ByteString and AsciiString hashCode algorithm uses all bits from each character. AsciiString caches the case insensitive hashCode result.