@Scottmitch 2015-04-25
@nmittler @buchgr - FYI.  There are a bunch of benchmarks here.  I will run and get results next week.  Note that the ascii string case insensitive is cached now so the benchmark many not be too useful for that case.