@Scottmitch 2015-04-27
@nmittler - Thanks for the review!  Turns out the performance results did not post so favorably for the case insensitive stuff.  Lets consider this a work in progress for the time being....I'm looking into alternatives and will ping you if I come up with anything.