@Scottmitch 2015-04-28
@nmittler - I reverted back to a [FNV-1a](http://en.wikipedia.org/wiki/Fowler%E2%80%93Noll%E2%80%93Vo_hash_function) inspired hash function.  The rational being performance degraded for the case insensitive case when attempting to tie case insensitivity into the Murmur32 algorithm.  One of the reasons why the FNV-1a influenced algorithm performs better is we can take advantage of `getLong` where Murmur32 only uses `getInt`.  I could have used a `Murmur` 64 bit version but murmur 64 is already more complicated, and incorporating case insensitive for byte[] and char[] would not pretty (the current state of this PR isn't exactly pretty either but conceptually "simple"...each step is `h =^ value; h *= PRIME` with the biggest size data we have left).

To get a baseline for how "well" the hash algorithms distribute hash codes I ran the old and new algorithms through a few scenarios (more can be done if necessary...for example random inputs may be good to do...):
- Around 60,000 [english lower case words](http://www.mieliestronk.com/wordlist.html).
  - Old had 65 buckets with collisions (max of 2 values that collided in any given bucket).
  - New had 1 bucket with a collision (2 values that collided).
- The header names from HttpHeaderNames.
  - Neither had any collisions (77 value sample size)
- The header values from HttpHeaderValues.
  - Neither had any collisions (42 value sample size)
- List ~300 of headers from [here](https://www.iana.org/assignments/message-headers/message-headers.txt)
  - Neither had any collisions.

The immediate benefits of this are that the case insensitive hash code is cached in `AsciiString`, and the regular byte[] (case sensitive) hash is faster in all cases.  The String and case insensitive hash is worse until about 10 characters long, and becomes better at around 14 characters long.  The longer term goal is to use hash code and equals for the basis of the `DefaultHeaders` map after @buchgr's improvement PR https://github.com/netty/netty/pull/3681 is pulled in.  The thought is that lookups based purely of hash codes and `equals` will provide the following benefits:
1. More efficient - Hash code computation can be cached in most cases (case insensitive for HTTP/1.x and regular for HTTP/2).
2. Remove reliance on comparators - Order is not required and is being used to facilitate case insensitive and case sensitive compatibility.  Comparisons for every insertions, lookup, and removal should not be necessary (and can not be reasonably cached).
3. HTTP/1.x may generally have short headers names but HTTP/2 may have longer header "names" (binary data).  The longer the names the greater the pain when values are not cached, and comparisons must be done.
4. The values are subject to hash code and equals for the contains, removal conditional on value, and other operations which require finding a key/value pair.

```
Benchmark                                                   (size)   Mode  Cnt         Score         Error  Units
PlatformDependentBenchmark.arraysBytesHashCode                  10  thrpt   20  53684243.704 ±  194110.664  ops/s
PlatformDependentBenchmark.arraysBytesHashCode                  30  thrpt   20  17925689.590 ±   66577.121  ops/s
PlatformDependentBenchmark.arraysBytesHashCode                  50  thrpt   20  10421145.217 ±   30033.326  ops/s
PlatformDependentBenchmark.arraysBytesHashCode                 100  thrpt   20   5061790.267 ±    5387.161  ops/s
PlatformDependentBenchmark.arraysBytesHashCode                1000  thrpt   20    498281.519 ±     834.745  ops/s
PlatformDependentBenchmark.arraysBytesHashCode               10000  thrpt   20     49793.572 ±      55.194  ops/s
PlatformDependentBenchmark.arraysBytesHashCode              100000  thrpt   20      4985.099 ±       2.327  ops/s
PlatformDependentBenchmark.oldHashBytes                         10  thrpt   20  48914959.267 ±   34678.514  ops/s
PlatformDependentBenchmark.oldHashBytes                         30  thrpt   20  17510039.707 ±   15032.515  ops/s
PlatformDependentBenchmark.oldHashBytes                         50  thrpt   20  10283178.612 ±    7353.302  ops/s
PlatformDependentBenchmark.oldHashBytes                        100  thrpt   20   5027150.897 ±    3377.766  ops/s
PlatformDependentBenchmark.oldHashBytes                       1000  thrpt   20    496539.115 ±     703.904  ops/s
PlatformDependentBenchmark.oldHashBytes                      10000  thrpt   20     49790.428 ±      37.325  ops/s
PlatformDependentBenchmark.oldHashBytes                     100000  thrpt   20      4976.901 ±       3.661  ops/s
PlatformDependentBenchmark.oldHashCaseInsensitiveBytes          10  thrpt   20  44368565.448 ±   82536.778  ops/s
PlatformDependentBenchmark.oldHashCaseInsensitiveBytes          30  thrpt   20  16995940.412 ±   39103.185  ops/s
PlatformDependentBenchmark.oldHashCaseInsensitiveBytes          50  thrpt   20  10150494.434 ±    6688.047  ops/s
PlatformDependentBenchmark.oldHashCaseInsensitiveBytes         100  thrpt   20   4888024.560 ±   58940.024  ops/s
PlatformDependentBenchmark.oldHashCaseInsensitiveBytes        1000  thrpt   20    467459.527 ±     411.452  ops/s
PlatformDependentBenchmark.oldHashCaseInsensitiveBytes       10000  thrpt   20     46754.046 ±      37.408  ops/s
PlatformDependentBenchmark.oldHashCaseInsensitiveBytes      100000  thrpt   20      4702.920 ±       4.195  ops/s
PlatformDependentBenchmark.oldHashCaseInsensitiveString         10  thrpt   20  36341480.324 ±  402872.191  ops/s
PlatformDependentBenchmark.oldHashCaseInsensitiveString         30  thrpt   20  15974117.643 ±   27522.390  ops/s
PlatformDependentBenchmark.oldHashCaseInsensitiveString         50  thrpt   20   9757928.288 ±   16078.794  ops/s
PlatformDependentBenchmark.oldHashCaseInsensitiveString        100  thrpt   20   4606635.667 ±   10040.857  ops/s
PlatformDependentBenchmark.oldHashCaseInsensitiveString       1000  thrpt   20    485581.659 ±     537.035  ops/s
PlatformDependentBenchmark.oldHashCaseInsensitiveString      10000  thrpt   20     49161.695 ±      44.940  ops/s
PlatformDependentBenchmark.oldHashCaseInsensitiveString     100000  thrpt   20      4819.364 ±       4.390  ops/s
PlatformDependentBenchmark.oldHashString                        10  thrpt   20  41880339.595 ± 2331134.350  ops/s
PlatformDependentBenchmark.oldHashString                        30  thrpt   20  16764608.993 ±   47491.730  ops/s
PlatformDependentBenchmark.oldHashString                        50  thrpt   20   9968462.446 ±  137085.223  ops/s
PlatformDependentBenchmark.oldHashString                       100  thrpt   20   4995068.024 ±    3436.517  ops/s
PlatformDependentBenchmark.oldHashString                      1000  thrpt   20    497498.988 ±     496.297  ops/s
PlatformDependentBenchmark.oldHashString                     10000  thrpt   20     49817.076 ±      31.989  ops/s
PlatformDependentBenchmark.oldHashString                    100000  thrpt   20      4981.925 ±       3.025  ops/s
PlatformDependentBenchmark.unsafeHashBytes                      10  thrpt   20  69992615.346 ±   79618.948  ops/s
PlatformDependentBenchmark.unsafeHashBytes                      30  thrpt   20  54822763.141 ±   59138.817  ops/s
PlatformDependentBenchmark.unsafeHashBytes                      50  thrpt   20  45889804.182 ±  468024.519  ops/s
PlatformDependentBenchmark.unsafeHashBytes                     100  thrpt   20  34101994.590 ±   47607.420  ops/s
PlatformDependentBenchmark.unsafeHashBytes                    1000  thrpt   20   3985239.052 ±    5104.231  ops/s
PlatformDependentBenchmark.unsafeHashBytes                   10000  thrpt   20    398524.548 ±     463.329  ops/s
PlatformDependentBenchmark.unsafeHashBytes                  100000  thrpt   20     39839.092 ±      33.170  ops/s
PlatformDependentBenchmark.unsafeHashCaseInsensitiveBytes       10  thrpt   20  37223128.694 ±   73718.631  ops/s
PlatformDependentBenchmark.unsafeHashCaseInsensitiveBytes       30  thrpt   20  20096143.163 ±   30552.084  ops/s
PlatformDependentBenchmark.unsafeHashCaseInsensitiveBytes       50  thrpt   20  12481167.386 ±   32764.789  ops/s
PlatformDependentBenchmark.unsafeHashCaseInsensitiveBytes      100  thrpt   20   5676125.422 ±    6716.475  ops/s
PlatformDependentBenchmark.unsafeHashCaseInsensitiveBytes     1000  thrpt   20    582292.152 ±    5646.366  ops/s
PlatformDependentBenchmark.unsafeHashCaseInsensitiveBytes    10000  thrpt   20     57845.631 ±      90.937  ops/s
PlatformDependentBenchmark.unsafeHashCaseInsensitiveBytes   100000  thrpt   20      5753.999 ±      16.067  ops/s
PlatformDependentBenchmark.unsafeHashCaseInsensitiveString      10  thrpt   20  34520048.631 ±   18953.931  ops/s
PlatformDependentBenchmark.unsafeHashCaseInsensitiveString      30  thrpt   20  18458987.044 ±   33189.434  ops/s
PlatformDependentBenchmark.unsafeHashCaseInsensitiveString      50  thrpt   20  12902893.314 ±   94205.409  ops/s
PlatformDependentBenchmark.unsafeHashCaseInsensitiveString     100  thrpt   20   5258875.359 ±   72814.684  ops/s
PlatformDependentBenchmark.unsafeHashCaseInsensitiveString    1000  thrpt   20    563936.040 ±    9961.020  ops/s
PlatformDependentBenchmark.unsafeHashCaseInsensitiveString   10000  thrpt   20     57621.202 ±     157.593  ops/s
PlatformDependentBenchmark.unsafeHashCaseInsensitiveString  100000  thrpt   20      5728.587 ±      19.888  ops/s
PlatformDependentBenchmark.unsafeHashString                     10  thrpt   20  40772628.266 ±  158282.559  ops/s
PlatformDependentBenchmark.unsafeHashString                     30  thrpt   20  23261245.709 ±  162753.206  ops/s
PlatformDependentBenchmark.unsafeHashString                     50  thrpt   20  17569253.867 ±   16731.727  ops/s
PlatformDependentBenchmark.unsafeHashString                    100  thrpt   20  10794132.542 ±    4677.632  ops/s
PlatformDependentBenchmark.unsafeHashString                   1000  thrpt   20   1257363.971 ±     895.875  ops/s
PlatformDependentBenchmark.unsafeHashString                  10000  thrpt   20    129841.491 ±     170.427  ops/s
PlatformDependentBenchmark.unsafeHashString                 100000  thrpt   20     13025.864 ±       8.407  ops/s
```